+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 0.2.1
 
 * Fixed errors in dependency versions so plug-in works on versions of Eclipse back to 3.6.2.
+* Upgraded the Splunk SDK for Java version 1.2.1.
 	
 ## Version 0.2